twister: move cmake cache class
Move cmake cache classes into own file.
